feat: pontos aceitam multiplos de 0.5 (meio ponto)
- Schema: products.points e sale_items.points_each viram numeric(10,2)
  com mode: number no Drizzle (continua tipado como number em JS)
- Migration 0004 aplica ALTER TYPE numeric(10,2) preservando dados
- Validator productSchema valida multiplos de 0.5 com refine
- SUMs envolvendo points_each agora cast pra ::float8 (em vez de ::int)
  pra preservar decimais
- Inputs de produto com step=0.5 e texto auxiliar
- Helper formatPoints(value) em format.ts pra display BR (50.5 -> "50,5")

+++ src/app/admin/(app)/vendas/page.tsx
@@ -124,7 +124,7 @@ export default async function AdminSalesPage({
       storeId: sales.storeId,
       clerkName: clerks.name,
       storeName: stores.name,
-      points: sql<number>`COALESCE(SUM(${saleItems.quantity} * ${saleItems.pointsEach}), 0)::int`,
+      points: sql<number>`COALESCE(SUM(${saleItems.quantity} * ${saleItems.pointsEach}), 0)::float8`,
       items: sql<number>`COALESCE(SUM(${saleItems.quantity}), 0)::int`,
     })
     .from(sales)
